Refactor and remove redundant DTOs, update mappings and usages
Removed unused DTOs across the project to streamline the codebase and reduce redundancy. Updated related GraphQL and backend service mappings to align with this change, standardizing DTO usage and structure. Adjusted affected test cases and GraphQL queries to maintain functionality.

+++ src/main/java/org/ecommerce/common/dto/OrderSummaryDto.java
@@ -4,6 +4,8 @@ import lombok.Getter;
 import lombok.Setter;
 import org.eclipse.microprofile.graphql.Type;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Type
@@ -13,5 +15,5 @@ public class OrderSummaryDto
     private String orderDate;        // ISO-8601 string
     private String status;           // OrderStatusEn.name()
     private int itemCount;           // sum of quantities across all line items
-    private double totalAmount;      // order total as double
+    private BigDecimal totalAmount;
 }
